# UbyTecAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "UbyTecAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/TiposComercio": {
      "get": {
        "tags": [
          "TiposComercio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TipoComercio"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TiposComercio"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TipoComercio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/TipoComercio"
            }
          }
        }
      }
    },
...
    "/api/TiposComercio/{id}": {
      "get": {
        "tags": [
          "TiposComercio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/TipoComercio"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TiposComercio"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TipoComercio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TiposComercio"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TiposComercio/eliminarPorNombre/{nombre}": {
      "delete": {
        "tags": [
          "TiposComercio"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "nombre",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Afiliado": {
      "type": "object",
      "properties": {
        "afiliadoID": {
          "format": "int32",
          "type": "integer"
        },
        "numeroCedulaJuridica": {
          "type": "string"
        },
        "nombreComercio": {
          "type": "string"
        },
        "tipoComercio": {
          "type": "string"
        },
        "direccionProvincia": {
          "type": "string"
        },
        "direccionCanton": {
          "type": "string"
        },
        "direccionDistrito": {
          "type": "string"
        },
        "correoElectronico": {
          "type": "string"
        },
        "numeroSINPE": {
          "type": "string"
        },
        "administrador": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        },
        "telefonos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TelefonoAfiliado"
          }
        },
        "productos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Producto"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Producto": {
      "type": "object",
      "properties": {
        "productoID": {
          "format": "int32",
          "type": "integer"
        },
        "afiliadoID": {
          "format": "int32",
          "type": "integer"
        },
        "nombreProducto": {
          "type": "string"
        },
        "categoria": {
          "type": "string"
        },
        "foto": {
          "type": "string"
        },
        "precio": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "TelefonoAfiliado": {
      "type": "object",
      "properties": {
        "telefonoAfiliadoID": {
          "format": "int32",
          "type": "integer"
        },
        "afiliadoID": {
          "format": "int32",
          "type": "integer"
        },
        "numeroTelefono": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "TipoComercio": {
      "type": "object",
      "properties": {
        "tipoComercioID": {
          "format": "int32",
          "type": "integer"
        },
        "nombreTipo": {
          "type": "string"
        },
        "afiliados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Afiliado"
          }
        }
      },
      "additionalProperties": false
    },
...
  }
}